# Reduce allocations from the timeout handler by pooling the timers