<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

As per title, nothing more to say.

**Note to reviewers:** Please keep me honest on the invariants of `Stop` returning false and thus us properly draining the channel.

## Benchmark results

```
benchmark                                                old ns/op     new ns/op     delta
BenchmarkTimeToFirstByteTimeoutHandler/sequential-16     2023          1843          -8.90%
BenchmarkTimeToFirstByteTimeoutHandler/parallel-16       148           133           -10.47%

benchmark                                                old allocs     new allocs     delta
BenchmarkTimeToFirstByteTimeoutHandler/sequential-16     9              6              -33.33%
BenchmarkTimeToFirstByteTimeoutHandler/parallel-16       9              6              -33.33%

benchmark                                                old bytes     new bytes     delta
BenchmarkTimeToFirstByteTimeoutHandler/sequential-16     830           634           -23.61%
BenchmarkTimeToFirstByteTimeoutHandler/parallel-16       843           639           -24.20%
```

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @julz